feat(vector-def): add flattened indices for nested vectors
For an expression `c(1, 2, c(3, 4))` the nested vector would be flattened so that it equals `c(1, 2, 3, 4)`.
For this to work, we need to fetch the indices for the nested vector and add them to the existing indices.
In a case where there are values after a nested index we need to rewrite the indices so that they are in correct order.
E.g. for `c(1, c(2, 3), 4)` the argument with value '4' has initially the index 3, after it has been flattened,
it has to be 4. This is done by merging both lists with new indices similar to the merging part of MergeSort.