feat: process indices on the way up the fold
On the way down, we don't have all information e.g. about nested containers, which makes it inevitable to resolve them on the way up.
Previously, for a vector definition nested containers were handled on the way up, while primitves were handled on the way down.
This lead to a rather complex merging logic. Handling now everything on the way up let's us iterate the arguments once and define the
indices for the container at the same time. While adding the resolving of unnamed list arguments, I changed this too for list definitions.